+++ packages/react-await/src/utils/hashKey.ts
@@ -2,7 +2,7 @@ import type { Key } from '../Await'
 import { type PlainObject, isPlainObject } from './isPlainObject'
 
 export const hashKey = (key: Key) =>
-  JSON.stringify(key, (_, val) =>
+  JSON.stringify(key, (_, val: unknown) =>
     isPlainObject(val)
       ? Object.keys(val)
           .sort()