+++ packages/react-query/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @suspensive/react-query
 
+## 1.22.1
+
+### Patch Changes
+
+- 3782ae8: ci(eslint): add plugin:@typescript-eslint/strict-type-checked
+- Updated dependencies [3782ae8]
+  - @suspensive/react@1.22.1
+
 ## 1.22.0
 
 ### Minor Changes